refactor(games-list): extract launch-flow into a part file
Move the game-launch orchestration slice of `_SystemGamesListState` verbatim
into `my_games_list/launch_flow.dart` as `extension _LaunchFlow`. Methods:
`_freeMemoryForGameplay`, `_detectGameSavesForSelectedGame`,
`_resolveSystemFolderName`, `_pushAchievementsForLaunch`,
`_reactivateGamepadNavigation`, `_selectCurrentGame`, `_showRandomGameDialog`.

Pure move: bodies are byte-identical to `main` (token-stream verified) apart
from the two mechanical (a)-split substitutions — `setState(...)` routed
through the host `rebuild()` bridge (`State.setState` is `@protected`), and the
static `_log` qualified as `_SystemGamesListState._log`. All state stays on the
host. Host 2290 -> 1733; new file 559 lines.
